+++ pkg/service/service_test.go
@@ -1,8 +1,8 @@
-package core_test
+package service_test
 
 import (
-	"bitcoinrateapp/pkg/core"
 	"bitcoinrateapp/pkg/model"
+	"bitcoinrateapp/pkg/service"
 	"bitcoinrateapp/pkg/testenv"
 	"errors"
 	"testing"
@@ -10,8 +10,8 @@ import (
 
 func TestServiceRate(t *testing.T) {
 	rate := 100.0
-	service := core.NewService(nil, &testenv.MockRate{ExpectedRate: rate}, nil)
-	actualRate, err := service.ExchangeRate()
+	btcservice := service.NewService(nil, &testenv.MockRate{ExpectedRate: rate}, nil)
+	actualRate, err := btcservice.ExchangeRate()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -27,9 +27,9 @@ func TestServiceSubscribeSuccessfully(t *testing.T) {
 		t.Fatal(err)
 	}
 	db := &testenv.MockDB{}
-	service := core.NewService(db, nil, nil)
+	btcservice := service.NewService(db, nil, nil)
 
-	err = service.Subscribe(subscriber)
+	err = btcservice.Subscribe(subscriber)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -47,11 +47,11 @@ func TestServiceSubscribeError(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	expError := core.ErrIsDuplicate
+	expError := service.ErrIsDuplicate
 	db := &testenv.MockErrorDB{ExpectedError: expError}
-	service := core.NewService(db, nil, nil)
+	btcservice := service.NewService(db, nil, nil)
 
-	err = service.Subscribe(subscriber)
+	err = btcservice.Subscribe(subscriber)
 	if !errors.Is(err, expError) {
 		t.Fatal(err)
 	}
@@ -63,9 +63,9 @@ func TestServiceNotify(t *testing.T) {
 	db := &testenv.MockDB{Memory: receivers}
 	rateProvider := &testenv.MockRate{ExpectedRate: rate}
 	sender := &testenv.MockSender{}
-	service := core.NewService(db, rateProvider, sender)
+	btcservice := service.NewService(db, rateProvider, sender)
 
-	err := service.Notify()
+	err := btcservice.Notify()
 
 	if err != nil {
 		t.Fatal(err)
