+++ pkg/storage/filedb.go
@@ -1,4 +1,4 @@
-package core
+package storage
 
 import (
 	"bufio"
