+++ test/functional/server_test.go
@@ -2,8 +2,8 @@ package functional_test
 
 import (
 	"bitcoinrateapp/pkg/app"
-	"bitcoinrateapp/pkg/core"
 	"bitcoinrateapp/pkg/email"
+	"bitcoinrateapp/pkg/service"
 	"bitcoinrateapp/pkg/testenv"
 	"errors"
 	"fmt"
@@ -36,8 +36,8 @@ func TestHTTPServer(t *testing.T) {
 	sender := email.NewSender(client, formatter)
 	rateRequester := &testenv.MockRate{ExpectedRate: 1000}
 
-	service := core.NewService(receivers, rateRequester, sender)
-	handler := app.NewExchangeRateHandler(service)
+	btcservice := service.NewService(receivers, rateRequester, sender)
+	handler := app.NewExchangeRateHandler(btcservice)
 	addr := "localhost:3333"
 	startServer(handler, addr, t)
 
@@ -107,7 +107,7 @@ func testSubscribe(addr string, t *testing.T) {
 	}
 }
 
-func testDuplicateSubscribe(addr string, db core.Storage[string], t *testing.T) {
+func testDuplicateSubscribe(addr string, db service.Storage[string], t *testing.T) {
 	address := "test@test"
 	err := db.Append(address)
 	if err != nil {
